requirements.txt
streamlit
langchain-groq
langgraph
langchain-core
python-dotenv
langgraph-checkpoint-sqlite
transformers
fastapi
uvicorn
requests
langgraph-checkpoint-postgres
psycopg[binary] 
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart
pydantic[email]